+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-12T06:27:06.026Z",
+    "2026-06-12T07:58:13.213Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {