+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -182,7 +182,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KTX87DJV779XXVR1V6HJYDP2"),
+  _id: z.string().default("01KTXDE762JD0X71PATXD9YBTS"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: ListDatasourcesStatus$inboundSchema,